gx/pyproject.toml
[project]
name = "gx-project"
version = "0.1.0"
description = "Great Expectations project for data quality checks"
requires-python = ">=3.8"
dependencies = [
    "great-expectations[snowflake]>=0.16.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
